Do not delete working directories when debug is enabled.
Contents of partest's working directories is very useful for
debugging. Do not delete it in case od --debug option
enabled.

More details here:
https://groups.google.com/d/topic/scala-internals/rGNeYIQTL3s/discussion

+++ src/partest/scala/tools/partest/nest/Worker.scala
@@ -150,7 +150,8 @@ class Worker(val fileManager: FileManager, params: TestRunParams) extends Actor 
   private def updateStatus(key: String, num: Int) = status(key) = num
   
   private def cleanup() {
-    toDelete foreach (_.deleteRecursively())
+    if (!scala.tools.partest.isPartestDebug)
+      toDelete foreach (_.deleteRecursively())
     toDelete.clear()
   }
   sys addShutdownHook cleanup()
